backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.26.0
requests==2.31.0
openai==1.10.0
langchain==0.1.4
langchain-openai==0.0.5
websockets==12.0
jira==3.5.2
pygithub==2.1.1
atlassian-python-api==3.41.0
bcrypt==3.2.0
pgvector==0.3.6
openpyxl==3.1.2
python-docx==1.1.0
PyYAML==6.0.1

